requirements.txt
python-dotenv
aiogram
tensorflow
pandas
scikit-learn
emoji